+++ include/rizz/graphics.h
@@ -384,6 +384,11 @@ typedef struct rizz_api_gfx {
     sg_features (*query_features)(void);
     sg_limits (*query_limits)(void);
     sg_pixelformat_info (*query_pixelformat)(sg_pixel_format fmt);
+    sg_buffer_desc (*query_buffer_defaults)(const sg_buffer_desc* desc);
+    sg_image_desc (*query_image_defaults)(const sg_image_desc* desc);
+    sg_shader_desc (*query_shader_defaults)(const sg_shader_desc* desc);
+    sg_pipeline_desc (*query_pipeline_defaults)(const sg_pipeline_desc* desc);
+    sg_pass_desc (*query_pass_defaults)(const sg_pass_desc* desc);
 
     // internal use (imgui plugin)
     void (*internal_state)(void** make_cmdbuff, int* make_cmdbuff_sz);